refactor/serializer file structure
