+++ src/openpecha/pecha/serializers/pecha_db/translation.py
