+++ tests/pecha/serializers/pecha_db/commentary/complex/test_complex_commentary.py
@@ -2,7 +2,7 @@ from pathlib import Path
 from unittest import TestCase, mock
 
 from openpecha.pecha import Pecha
-from openpecha.pecha.serializers.commentary.complex_commentary import (
+from openpecha.pecha.serializers.pecha_db.commentary.complex_commentary import (
     ComplexCommentarySerializer,
 )
 from openpecha.utils import read_json
@@ -44,7 +44,7 @@ class TestCommentarySerializer(TestCase):
     def setUp(self):
         # Create the patcher and set return_value
         self.patcher = mock.patch(
-            "openpecha.pecha.serializers.commentary.complex_commentary.CategoryExtractor.get_category",
+            "openpecha.pecha.serializers.pecha_db.commentary.complex_commentary.CategoryExtractor.get_category",
             return_value={
                 "bo": [
                     {"name": "སངས་རྒྱས་ཀྱི་བཀའ།", "heDesc": "", "heShortDesc": ""},
@@ -64,7 +64,7 @@ class TestCommentarySerializer(TestCase):
         self.mock_get_category = self.patcher.start()
 
     @mock.patch(
-        "openpecha.pecha.serializers.commentary.complex_commentary.get_en_content_translation",
+        "openpecha.pecha.serializers.pecha_db.commentary.complex_commentary.get_en_content_translation",
         return_value=MOCK_BO_TO_EN_TRANSLATION,
     )
     def test_bo_commentary_serializer(self, mock_get_en_translation):
@@ -77,7 +77,7 @@ class TestCommentarySerializer(TestCase):
         assert serialized_json == expected_serialized_json
 
     @mock.patch(
-        "openpecha.pecha.serializers.commentary.complex_commentary.get_bo_content_translation",
+        "openpecha.pecha.serializers.pecha_db.commentary.complex_commentary.get_bo_content_translation",
         return_value=MOCK_EN_TO_BO_TRANSLATION,
     )
     def test_en_commentary_serializer(self, mock_get_bo_translation):
@@ -90,7 +90,7 @@ class TestCommentarySerializer(TestCase):
         assert serialized_json == expected_serialized_json
 
     @mock.patch(
-        "openpecha.pecha.serializers.commentary.complex_commentary.get_bo_content_translation",
+        "openpecha.pecha.serializers.pecha_db.commentary.complex_commentary.get_bo_content_translation",
         return_value=MOCK_ZH_TO_BO_TRANSLATION,
     )
     def test_zh_commentary_serializer(self, mock_get_bo_translation):
