+++ tests/pecha/serializers/pecha_db/translation/test_translation_serializer.py
@@ -2,7 +2,7 @@ from pathlib import Path
 from unittest import TestCase, mock
 
 from openpecha.pecha import Pecha
-from openpecha.pecha.serializers.translation import TextTranslationSerializer
+from openpecha.pecha.serializers.pecha_db.translation import TextTranslationSerializer
 from openpecha.utils import read_json
 
 DATA_DIR = Path(__file__).parent / "data"
@@ -12,7 +12,7 @@ class TestTextTranslationSerializer(TestCase):
     def setUp(self):
         # Create the patcher and set return_value
         self.patcher = mock.patch(
-            "openpecha.pecha.serializers.translation.CategoryExtractor.get_category",
+            "openpecha.pecha.serializers.pecha_db.translation.CategoryExtractor.get_category",
             return_value={
                 "bo": [
                     {"name": "སངས་རྒྱས་ཀྱི་བཀའ།", "heDesc": "", "heShortDesc": ""},
@@ -53,7 +53,7 @@ class TestTextTranslationSerializer(TestCase):
         root_pecha = Pecha.from_path(root_opf)
         translation_pecha = Pecha.from_path(translation_opf)
         with mock.patch(
-            "openpecha.pecha.serializers.translation.get_pecha_with_id"
+            "openpecha.pecha.serializers.pecha_db.translation.get_pecha_with_id"
         ) as mock_get_pecha_with_id:
             mock_get_pecha_with_id.return_value = root_pecha
